add client to db

+++ src/main/java/com/app/dataentry/repositories/ClientRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.app.dataentry.model.Client;
 
 @Repository
-public interface ClientRepository extends PagingAndSortingRepository<Client, Integer> {
+public interface ClientRepository extends PagingAndSortingRepository<Client, Long> {
 
 }
